+++ projects/start-wrt/web/src/app/services/api/api.service.ts
@@ -583,7 +583,6 @@ export type WanDnsSetRequest = {
 }
 
 export type WanDdnsProvider =
-  | 'start9'
   | 'dyndns'
   | 'noip'
   | 'cloudflare'